Complete codebase restructuring with file moves, removals, and configuration cleanup
