--- demo_semantic_cache_fix.py
@@ -1,231 +1,0 @@
-#!/usr/bin/env python3
-"""
-Demonstration script showing the semantic cache fix in action.
-This script validates that unrelated queries no longer get false positive cache hits.
-"""
-
-import asyncio
-from unittest.mock import Mock, AsyncMock
-from agentic_rag.app.semantic_cache import SemanticCache
-
-
-def create_demo_cache():
-    """Create a semantic cache instance with demo configuration."""
-    cache = SemanticCache()
-    
-    # Mock dependencies for demonstration
-    cache.redis_client = AsyncMock()
-    cache.weaviate_client = Mock()
-    cache.cache_vector_store = Mock()
-    cache.embedding_model = Mock()
-    cache._initialized = True
-    
-    return cache
-
-
-async def demo_original_bug_scenario():
-    """Demonstrate the original bug scenario and how it's now fixed."""
-    print("🔍 DEMONSTRATING SEMANTIC CACHE FIX")
-    print("=" * 50)
-    
-    cache = create_demo_cache()
-    
-    # Simulate the original bug scenario
-    print("📝 Original Bug Scenario:")
-    print("  Query 1 (cached): 'What is Nativepath??'")
-    print("  Query 2 (test):   'What are the benefits of consuming MCTs daily?'")
-    print()
-    
-    # Mock the cached Nativepath entry
-    nativepath_entry = {
-        "query": "What is Nativepath??",
-        "answer": "NativePath is a company that offers health products, including MCT Oil Powder...",
-        "cache_id": "cache_nativepath_001"
-    }
-    
-    # Mock no exact match (different queries)
-    cache.redis_client.get.return_value = None
-    
-    # Mock vector similarity search returning the Nativepath query
-    mock_doc = Mock()
-    mock_doc.page_content = nativepath_entry["query"]
-    mock_doc.metadata = {"cache_id": nativepath_entry["cache_id"]}
-    
-    # Simulate the problematic high similarity score that caused the bug
-    cache._vector_similarity_search = AsyncMock(return_value=[(mock_doc, 1.0)])
-    
-    # Mock realistic similarity scores for unrelated queries (low values)
-    cache._embedding_similarity = Mock(return_value=0.22)  # Low embedding similarity
-    cache._ce_similarity = AsyncMock(return_value=0.15)    # Low cross-encoder similarity
-    cache._lexical_similarity = Mock(return_value=0.08)    # Very low lexical similarity
-    
-    # Mock cache retrieval
-    async def mock_get_cache_entry_by_id(cache_id):
-        if cache_id == nativepath_entry["cache_id"]:
-            return nativepath_entry
-        return None
-    cache._get_cache_entry_by_id = mock_get_cache_entry_by_id
-    
-    # Test the problematic query
-    mct_query = "What are the benefits of consuming MCTs daily?"
-    result = await cache.get_cached_answer(mct_query)
-    
-    print("🔧 Fix Applied:")
-    print("  - Added SEMANTIC_CACHE_SCORE_MODE='distance' for Weaviate compatibility")
-    print("  - Enhanced validation for suspiciously high similarity scores")
-    print("  - Multi-metric validation (vector + embedding + lexical similarity)")
-    print()
-    
-    print("✅ Result:")
-    if result is None:
-        print("  ✅ CACHE MISS - Unrelated queries correctly rejected!")
-        print("  ✅ The bug is FIXED - no false positive cache hit")
-    else:
-        print("  ❌ CACHE HIT - Bug still present!")
-        print(f"  ❌ Wrong answer returned: {result.get('answer', '')[:50]}...")
-    
-    return result is None
-
-
-async def demo_legitimate_similar_queries():
-    """Demonstrate that legitimate similar queries still get cache hits."""
-    print()
-    print("📋 TESTING LEGITIMATE SIMILAR QUERIES")
-    print("=" * 50)
-    
-    cache = create_demo_cache()
-    
-    # Original cached query about MCT benefits
-    original_query = "What are the benefits of consuming MCT daily?"
-    original_entry = {
-        "query": original_query,
-        "answer": "MCT benefits include improved energy metabolism, enhanced cognitive function...",
-        "cache_id": "cache_mct_001"
-    }
-    
-    # Mock no exact match
-    cache.redis_client.get.return_value = None
-    
-    # Mock vector similarity search returning the original query
-    mock_doc = Mock()
-    mock_doc.page_content = original_query
-    mock_doc.metadata = {"cache_id": original_entry["cache_id"]}
-    
-    # Mock cache retrieval
-    async def mock_get_cache_entry_by_id(cache_id):
-        if cache_id == original_entry["cache_id"]:
-            return original_entry
-        return None
-    cache._get_cache_entry_by_id = mock_get_cache_entry_by_id
-    cache._update_cache_access = AsyncMock(return_value=original_entry)
-    
-    # Test similar queries as specified in requirements
-    similar_queries = [
-        "Explain the benefits of MCT",
-        "What are the benefits of MCTs?", 
-        "How consuming MCT can be beneficial to the health?",
-        "What benefits can I expect if I consume MCT daily?"
-    ]
-    
-    print(f"📝 Original cached query: '{original_query}'")
-    print()
-    print("🔍 Testing similar queries that SHOULD get cache hits:")
-    
-    all_passed = True
-    for i, similar_query in enumerate(similar_queries, 1):
-        # Mock high similarity scores for legitimate similar queries
-        cache._vector_similarity_search = AsyncMock(return_value=[(mock_doc, 0.94)])
-        cache._embedding_similarity = Mock(return_value=0.91)
-        cache._ce_similarity = AsyncMock(return_value=0.88)
-        cache._lexical_similarity = Mock(return_value=0.65)  # Good lexical overlap
-        
-        result = await cache.get_cached_answer(similar_query)
-        
-        if result is not None:
-            print(f"  {i}. ✅ '{similar_query}' → CACHE HIT (correct)")
-        else:
-            print(f"  {i}. ❌ '{similar_query}' → CACHE MISS (should have been hit)")
-            all_passed = False
-    
-    print()
-    if all_passed:
-        print("✅ All similar queries correctly received cache hits!")
-    else:
-        print("❌ Some similar queries were incorrectly rejected")
-    
-    return all_passed
-
-
-async def demo_edge_cases():
-    """Demonstrate edge case handling."""
-    print()
-    print("⚠️  TESTING EDGE CASES")
-    print("=" * 50)
-    
-    cache = create_demo_cache()
-    
-    # Mock no exact match
-    cache.redis_client.get.return_value = None
-    
-    # Mock a suspicious high similarity case
-    mock_doc = Mock()
-    mock_doc.page_content = "What is artificial intelligence?"
-    mock_doc.metadata = {"cache_id": "cache_ai_001"}
-    
-    print("🔍 Edge Case: Suspiciously high vector similarity with low semantic overlap")
-    print("  Cached query: 'What is artificial intelligence?'")
-    print("  Test query:   'What are the health benefits of chocolate?'")
-    print("  Vector sim:   0.995 (suspiciously high)")
-    print("  Embedding:    0.30 (low)")
-    print("  Lexical:      0.02 (very low)")
-    print()
-    
-    # Mock suspicious similarity pattern
-    cache._vector_similarity_search = AsyncMock(return_value=[(mock_doc, 0.995)])
-    cache._embedding_similarity = Mock(return_value=0.30)
-    cache._lexical_similarity = Mock(return_value=0.02)
-    
-    result = await cache.get_cached_answer("What are the health benefits of chocolate?")
-    
-    if result is None:
-        print("✅ Edge case handled correctly - suspicious similarity rejected!")
-    else:
-        print("❌ Edge case failed - suspicious similarity was accepted")
-    
-    return result is None
-
-
-async def main():
-    """Run the complete demonstration."""
-    print("🚀 SEMANTIC CACHE FIX DEMONSTRATION")
-    print("🐛 Fixing false positive cache hits for unrelated queries")
-    print()
-    
-    # Test original bug scenario
-    bug_fixed = await demo_original_bug_scenario()
-    
-    # Test legitimate similar queries
-    similar_queries_work = await demo_legitimate_similar_queries()
-    
-    # Test edge cases
-    edge_cases_work = await demo_edge_cases()
-    
-    # Final summary
-    print()
-    print("📊 SUMMARY")
-    print("=" * 50)
-    print(f"✅ Original bug fixed:           {'YES' if bug_fixed else 'NO'}")
-    print(f"✅ Similar queries still work:   {'YES' if similar_queries_work else 'NO'}")
-    print(f"✅ Edge cases handled:           {'YES' if edge_cases_work else 'NO'}")
-    print()
-    
-    if bug_fixed and similar_queries_work and edge_cases_work:
-        print("🎉 ALL TESTS PASSED - Semantic cache fix is working correctly!")
-        print("🔒 Unrelated queries are now properly rejected")
-        print("✨ Similar queries continue to get cache hits as expected")
-    else:
-        print("⚠️  Some issues detected - review the results above")
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
