+++ pyproject.toml
@@ -48,6 +48,7 @@ pytest-mock = "^3.14.1"
 fpdf = "^1.7.2"
 httpx = "^0.28.1"
 ruff = "^0.12.11"
+pytest-asyncio = "^1.2.0"
 
 [tool.poetry.scripts]
 ingest = "agentic_rag.scripts.run_ingestion:main"
@@ -59,6 +60,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
+asyncio_mode = "auto"
 filterwarnings = [
     "ignore:Protobuf gencode version.*:UserWarning",
 ]
