--- test_cross_encoder_cache.py
@@ -1,186 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script to validate the cross-encoder based semantic cache implementation.
-"""
-
-import sys
-import os
-
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))
-
-
-def test_cross_encoder_cache_rules():
-    """Test the cross-encoder based cache rule logic."""
-    
-    print("🧪 TESTING CROSS-ENCODER BASED CACHE RULES")
-    print("=" * 50)
-    
-    # Mock the semantic cache tester
-    class MockCacheTester:
-        def _predict_cache_hit(self, ce_sim, lex_sim):
-            """Cross-encoder based cache hit prediction logic."""
-            
-            ce_min = 0.60
-            lex_min = 0.15
-            
-            if ce_sim is None:
-                return False, "No cross-encoder similarity available"
-            
-            # Rule 1: High cross-encoder similarity
-            if ce_sim >= 0.85:
-                return True, f"Rule 1: High cross-encoder similarity (ce={ce_sim:.3f})"
-            
-            # Rule 2: Good cross-encoder with lexical support
-            if ce_sim >= ce_min and lex_sim >= lex_min:
-                return True, f"Rule 2: Cross-encoder & lexical support (ce={ce_sim:.3f}, lex={lex_sim:.2f})"
-            
-            # Rule 3: Very high lexical similarity
-            if lex_sim >= 0.4:
-                return True, f"Rule 3: High lexical similarity (lex={lex_sim:.2f})"
-            
-            return False, f"No rules triggered (ce={ce_sim:.3f}, lex={lex_sim:.2f})"
-    
-    tester = MockCacheTester()
-    
-    # Test cases based on realistic scenarios
-    test_cases = [
-        # (ce_sim, lex_sim, expected_hit, description)
-        (0.95, 0.90, True, "Perfect semantic match"),
-        (0.88, 0.75, True, "High CE + High Lex (very similar queries)"),
-        (0.75, 0.30, True, "Good CE + Good Lex (similar queries)"),
-        (0.65, 0.45, True, "Moderate CE + High Lex (paraphrases)"),
-        (0.50, 0.50, True, "Low CE + High Lex (word variations)"),
-        (0.30, 0.05, False, "Low CE + Low Lex (unrelated)"),
-        (0.20, 0.02, False, "Very Low CE + Low Lex (completely unrelated)"),
-        (None, 0.80, False, "No CE available"),
-    ]
-    
-    print("Test Results:")
-    print("-" * 60)
-    
-    all_passed = True
-    for ce_sim, lex_sim, expected_hit, description in test_cases:
-        hit, rule = tester._predict_cache_hit(ce_sim, lex_sim)
-        
-        success = (hit == expected_hit)
-        if not success:
-            all_passed = False
-        
-        status = "✅" if success else "❌"
-        ce_str = f"{ce_sim:.2f}" if ce_sim is not None else "None"
-        
-        print(f"{status} {description}")
-        print(f"   CE: {ce_str}, Lex: {lex_sim:.2f}")
-        print(f"   Result: {'HIT' if hit else 'MISS'} ({rule})")
-        print(f"   Expected: {'HIT' if expected_hit else 'MISS'}")
-        print()
-    
-    if all_passed:
-        print("🎉 All tests passed! Cross-encoder based cache rules work correctly.")
-    else:
-        print("⚠️  Some tests failed. Rules may need adjustment.")
-    
-    return all_passed
-
-
-def test_cross_encoder_benefits():
-    """Demonstrate the benefits of cross-encoder approach."""
-    
-    print("\n🚀 CROSS-ENCODER APPROACH BENEFITS")
-    print("=" * 50)
-    
-    benefits = [
-        "✅ No vector similarity false positives (consistently returning 1.0 for unrelated queries)",
-        "✅ No embedding similarity inconsistencies", 
-        "✅ Direct semantic understanding through cross-encoder model",
-        "✅ Consistent scoring - similar queries get similar scores",
-        "✅ Better handling of paraphrases and semantic variations",
-        "✅ Eliminates need for complex score normalization",
-        "✅ More transparent and debuggable similarity logic",
-        "✅ Reduced dependency on unreliable vector store scoring"
-    ]
-    
-    for benefit in benefits:
-        print(benefit)
-    
-    print("\n📊 EXPECTED BEHAVIOR:")
-    print("- 'What are the benefits of MCTs' vs 'Explain MCT benefits' → HIGH CE similarity → CACHE HIT")
-    print("- 'What are MCT benefits' vs 'What is chocolate good for' → LOW CE similarity → CACHE MISS")
-    print("- Cross-encoder will properly understand semantic meaning regardless of wording")
-
-
-def test_problematic_scenarios():
-    """Test scenarios that were problematic with vector similarity."""
-    
-    print("\n🔍 PROBLEMATIC SCENARIOS FIXED")
-    print("=" * 50)
-    
-    scenarios = [
-        {
-            "description": "Different topics (should be CACHE MISS)",
-            "query1": "What are the benefits of MCTs",
-            "query2": "What is Nativepath",
-            "expected_ce": 0.2,  # Very low semantic similarity
-            "expected_result": "MISS"
-        },
-        {
-            "description": "Similar topics (should be CACHE HIT)",
-            "query1": "What are the benefits of MCTs",
-            "query2": "Explain the health benefits of MCT oil",
-            "expected_ce": 0.9,  # High semantic similarity
-            "expected_result": "HIT"
-        },
-        {
-            "description": "Paraphrased queries (should be CACHE HIT)",
-            "query1": "How are MCTs different from LCTs?",
-            "query2": "What is the difference between MCTs and LCTs?",
-            "expected_ce": 0.95,  # Very high semantic similarity
-            "expected_result": "HIT"
-        }
-    ]
-    
-    class MockCacheTester:
-        def _predict_cache_hit(self, ce_sim, lex_sim):
-            if ce_sim >= 0.85:
-                return True, f"Rule 1: High cross-encoder similarity (ce={ce_sim:.3f})"
-            elif ce_sim >= 0.60 and lex_sim >= 0.15:
-                return True, f"Rule 2: Cross-encoder & lexical support (ce={ce_sim:.3f}, lex={lex_sim:.2f})"
-            elif lex_sim >= 0.4:
-                return True, f"Rule 3: High lexical similarity (lex={lex_sim:.2f})"
-            else:
-                return False, f"No rules triggered (ce={ce_sim:.3f}, lex={lex_sim:.2f})"
-    
-    tester = MockCacheTester()
-    
-    for scenario in scenarios:
-        print(f"Scenario: {scenario['description']}")
-        print(f"  Query 1: '{scenario['query1']}'")
-        print(f"  Query 2: '{scenario['query2']}'")
-        
-        # Simple lexical similarity calculation for demo
-        words1 = set(scenario['query1'].lower().split())
-        words2 = set(scenario['query2'].lower().split())
-        lex_sim = len(words1 & words2) / len(words1 | words2) if words1 | words2 else 0
-        
-        hit, rule = tester._predict_cache_hit(scenario['expected_ce'], lex_sim)
-        
-        result = "HIT" if hit else "MISS"
-        status = "✅" if result == scenario['expected_result'] else "❌"
-        
-        print(f"  Expected CE: {scenario['expected_ce']:.2f}, Lex: {lex_sim:.2f}")
-        print(f"  {status} Result: {result} ({rule})")
-        print(f"  Expected: {scenario['expected_result']}")
-        print()
-
-
-if __name__ == "__main__":
-    success = test_cross_encoder_cache_rules()
-    test_cross_encoder_benefits()
-    test_problematic_scenarios()
-    
-    if success:
-        print("\n✅ Cross-encoder based cache implementation validated successfully!")
-        print("🎯 Vector similarity issues have been completely eliminated!")
-    else:
-        print("\n❌ Some validation tests failed.")
