--- validate_fix.py
@@ -1,136 +1,0 @@
-#!/usr/bin/env python3
-"""
-Simple validation script to check threshold logic without dependencies.
-"""
-
-def test_threshold_logic():
-    """Test the threshold logic that was updated."""
-    print("🔍 Testing Updated Threshold Logic")
-    print("="*50)
-    
-    # Current configuration values
-    vec_accept = 0.92      # SEMANTIC_CACHE_VECTOR_ACCEPT
-    vec_min = 0.85         # SEMANTIC_CACHE_VECTOR_MIN
-    emb_min = 0.88         # SEMANTIC_CACHE_EMB_ACCEPT
-    ce_min = 0.60          # SEMANTIC_CACHE_CE_ACCEPT
-    lex_min = 0.15         # SEMANTIC_CACHE_LEXICAL_MIN
-    
-    print(f"Configuration:")
-    print(f"  vec_accept = {vec_accept}")
-    print(f"  vec_min = {vec_min}")
-    print(f"  emb_min = {emb_min}")
-    print(f"  ce_min = {ce_min}")
-    print(f"  lex_min = {lex_min}")
-    print()
-    
-    # Test cases based on user's scenarios
-    test_cases = [
-        {
-            "name": "Query 1: 'What are the benefits of MCTs' (original)",
-            "vector_sim": 1.0,      # Exact match
-            "emb_sim": 1.0,
-            "ce_sim": 1.0,
-            "lex_sim": 1.0,
-            "expected": "HIT (exact match)"
-        },
-        {
-            "name": "Query 3: 'What are the benefits of taking MCT'",
-            "vector_sim": 0.88,     # Above vec_min
-            "emb_sim": 0.85,        # Slightly below emb_min (0.88) but close
-            "ce_sim": 0.72,         # Well above ce_min (0.60)
-            "lex_sim": 0.65,        # High lexical overlap
-            "expected": "HIT (Rule 4: moderate vector + semantic support)"
-        },
-        {
-            "name": "Query 4: 'What benefits can I expect if I consume MCT daily?'",
-            "vector_sim": 0.87,     # Above vec_min
-            "emb_sim": 0.84,        # Below emb_min but Rule 4 should apply
-            "ce_sim": 0.68,         # Above ce_min + 0.05 (0.65)
-            "lex_sim": 0.45,        # Moderate lexical overlap
-            "expected": "HIT (Rule 4: moderate vector + good CE)"
-        },
-        {
-            "name": "Query 6: 'What are the benefits of taking MCTs?'",
-            "vector_sim": 0.90,     # Above vec_min + 0.05
-            "emb_sim": 0.86,        # Close to emb_min
-            "ce_sim": 0.70,         # Well above ce_min
-            "lex_sim": 0.75,        # High lexical overlap
-            "expected": "HIT (Rule 4: good vector + semantic support)"
-        },
-        {
-            "name": "Unrelated query: 'What is chocolate made of?'",
-            "vector_sim": 0.45,     # Low vector similarity
-            "emb_sim": 0.30,        # Low embedding similarity
-            "ce_sim": 0.25,         # Low cross-encoder similarity
-            "lex_sim": 0.05,        # Very low lexical overlap
-            "expected": "MISS (all metrics too low)"
-        }
-    ]
-    
-    def evaluate_rules(vector_sim, emb_sim, ce_sim, lex_sim):
-        """Simulate the rule evaluation logic."""
-        accept = False
-        reason = ""
-        
-        # Rule 1: very high vector similarity alone
-        if vector_sim >= vec_accept:
-            # Additional validation for perfect or near-perfect scores
-            if vector_sim >= 0.99:
-                # Only reject if BOTH embedding and lexical similarities are very low
-                if emb_sim < 0.5 and lex_sim < 0.05:
-                    return False, "Suspicious high similarity rejected"
-            accept = True
-            reason = f"Rule 1: vector>={vec_accept}"
-        else:
-            # Rule 2: require BOTH cross-encoder and embedding support with vector above minimum
-            if vector_sim >= vec_min and ce_sim >= ce_min and emb_sim >= emb_min:
-                accept = True
-                reason = f"Rule 2: vector>={vec_min} & ce>={ce_min} & emb>={emb_min}"
-            # Rule 3: tiny lexical support helps borderline cases
-            elif vector_sim >= vec_min and ce_sim >= ce_min and emb_sim >= (emb_min - 0.03) and lex_sim >= lex_min:
-                accept = True
-                reason = f"Rule 3: vector>={vec_min} & ce>={ce_min} & emb~ & lex>={lex_min}"
-            # Rule 4: more lenient rule for similar queries
-            elif vector_sim >= (vec_min + 0.02) and (emb_sim >= (emb_min - 0.05) or ce_sim >= (ce_min + 0.05)):
-                accept = True
-                reason = f"Rule 4: moderate vector>={vec_min + 0.02} & semantic support"
-        
-        return accept, reason
-    
-    print("Test Results:")
-    print("-" * 80)
-    
-    all_passed = True
-    for test_case in test_cases:
-        accept, reason = evaluate_rules(
-            test_case['vector_sim'], 
-            test_case['emb_sim'], 
-            test_case['ce_sim'], 
-            test_case['lex_sim']
-        )
-        
-        result = "HIT" if accept else "MISS"
-        expected_hit = "HIT" in test_case['expected']
-        actual_hit = accept
-        
-        success = expected_hit == actual_hit
-        if not success:
-            all_passed = False
-        
-        status = "✅" if success else "❌"
-        print(f"{status} {test_case['name']}")
-        print(f"   Scores: vec={test_case['vector_sim']:.2f}, emb={test_case['emb_sim']:.2f}, ce={test_case['ce_sim']:.2f}, lex={test_case['lex_sim']:.2f}")
-        print(f"   Result: {result} ({reason})")
-        print(f"   Expected: {test_case['expected']}")
-        print()
-    
-    if all_passed:
-        print("🎉 All test cases passed! The updated thresholds should work correctly.")
-    else:
-        print("⚠️  Some test cases failed. Further adjustments may be needed.")
-    
-    return all_passed
-
-
-if __name__ == "__main__":
-    test_threshold_logic()
